boxMate/requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.6
django-filter==2.3.0
django-import-export==2.3.0
django-js-asset==1.2.2
django-model-utils==3.2.0
django-multiselectfield==0.1.12
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jdcal==1.4.1
jsonfield==2.0.2
lazy-object-proxy==1.4.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.6
pkg-resources==0.0.0
progressbar2==3.47.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-utils==2.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typed-ast==1.4.2
urllib3==1.25.11
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
